implementation/requirements.txt
numpy>=1.24.0
Pillow>=10.0.0
PyQt6>=6.5.0
psutil>=5.9.0